Extract testHelper.
